+++ src/components/screens/SwapScreen/hooks/index.ts
@@ -1,3 +1,3 @@
 // SwapScreen-specific hooks - simplified and professional
-export { useSwapReview } from "./useSwapReview";
 export { useTransactionStatus } from "./useTransactionStatus";
+export { useSwapPathFinding } from "./useSwapPathFinding";
